Update README.md
The dataset that I will be examining is the Mushroom Data Set from the UCI machine learning repository. This dataset is derived from the Audubon Society North American mushrooms field guide where mushrooms are described in terms of 22 physical characteristics (features) and their classifications as edible or poisonous. If the status of the mushrooms was of unknown edibility or not recommended for consumption, it was added to the poisonous category in the mushroom dataset. This classification problem is the one that I will be trying to obtain an accurate machine learning model to solve. This problem is interesting as there is no single mushroom attribute that determines the edibility of the fungus; the combination of features is what predicates the classification. The classification of mushrooms based on their characteristics could contribute to understanding the edibility of some 11,000 species of mushrooms native to North America and keep people from consuming toxic fungi.
For this project, I would like to explore different machine learning models for classification including Support Vector Machine (sklearn’s SVM function – preexisting), k nearest neighbors (KNN - assumes similar data points will be ‘near’ each other so may not work well for this problem), and fully connected neural networks from ‘scratch.’ Determining a model to best predict the edible/poisonous nature of the mushrooms presented is the main objective. This will include the optimization of each model via the tuning of hyper-parameters, etc. and then a thorough comparison of the models’ results. 
For results comparison, classification accuracy, precision, recall, and f1 scores will be used. These results will be tabulated to allow for clear and direct comparison between the models that are tested. Additionally, a confusion matrix may be used to visualize the effectiveness of each machine learning model’s ability to classify the mushrooms.